backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-dotenv==1.0.1
pydantic==2.9.2

# Database
supabase==2.9.0

# Google Cloud services
google-cloud-speech==2.28.0
google-cloud-translate==3.17.0
google-generativeai==0.8.3

# ML and AI
sentence-transformers==3.2.1
openai==1.54.4

# Audio processing for emotion analyzer
numpy>=1.24.3
librosa==0.11.0
soundfile==0.12.1
pydub==0.25.1
